add linmsg_dp

+++ include/rcm2rgrid.h
@@ -10,7 +10,9 @@
 
 namespace NCL_cxx {
 
-void rgrid2rcm(const double* const xi, const double* const yi, int xiSize, int yiSize, const double* const xo, const double* const yo, int xoSize, int yoSize, const double* const fi, float* const fo, double msgValue, int &ier);
+
+
+void rgrid2rcm(const double* const xi, const double* const yi, int xiSize, int yiSize, const double* const xo, const double* const yo, int xoSize, int yoSize, const double* const fi, double* const fo, double msgValue, int &ier);
 
 }
 #endif /* rcm2rgrid_h */
